+++ Carbon/Functions/Clear-DscLocalResourceCache.ps1
@@ -21,9 +21,9 @@ function Clear-CDscLocalResourceCache
 
     Use-CallerPreference -Cmdlet $PSCmdlet -Session $ExecutionContext.SessionState
 
-    Get-Cim -Class 'msft_providers' | 
-    Where-Object {$_.provider -like 'dsccore'} | 
-    Select-Object -ExpandProperty HostProcessIdentifier | 
-    ForEach-Object { Get-Process -ID $_ } | 
-    Stop-Process -Force
+    Get-CCimInstance -Class 'msft_providers' | 
+        Where-Object {$_.provider -like 'dsccore'} |
+        Select-Object -ExpandProperty HostProcessIdentifier |
+        ForEach-Object { Get-Process -ID $_ } |
+        Stop-Process -Force
 }