+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 ## Added suport for PowerShell Core
 
-* Added: `Get-Cim` function that determines which method to call when accessing a computer's Common Information Model based on the version of PowerShell being used. `Get-Cim` will call `Get-CimInstance` under PowerShell Core and `Get-WmiObject` under Windows PowerShell. `Get-WmiObject` has been deprecated under PowerShell Core.
+* Added: `Get-CCimInstance` function that determines which method to call when accessing a computer's Common Information Model based on the version of PowerShell being used. `Get-CCimInstance` will call `Get-CimInstance` under PowerShell Core and `Get-WmiObject` under Windows PowerShell. `Get-WmiObject` has been deprecated under PowerShell Core.
 * Fixed: Carbon will now use version 2 of the `PSDesiredStateConfiguration` module when running under PowerShell Core as verion 1 was incompatible.
 * Fixed: Granting permissions on a certificate will use type `Security.AccessControl.FileSystemRights` as `Security.AccessControl.CryptoKeyRights` has been deprecated under PowerShell Core.